refactor: Splitting Subscription tests into multiple files.
